add geo_h3.rs
